+++ scim2_models/rfc7643/schema.py
@@ -5,6 +5,7 @@ from typing import Optional
 
 from ..base import ComplexAttribute
 from ..base import Mutability
+from ..base import Required
 from ..base import Returned
 from ..base import Uniqueness
 from .resource import Meta
@@ -23,23 +24,23 @@ class Attribute(ComplexAttribute):
         binary = "binary"
         complex = "complex"
 
-    name: Annotated[str, Mutability.read_only]
+    name: Annotated[str, Mutability.read_only, Required.true]
     """The attribute's name."""
 
-    type: Annotated[Type, Mutability.read_only]
+    type: Annotated[Type, Mutability.read_only, Required.true]
     """The attribute's data type."""
 
     sub_attributes: Annotated[Optional[List["Attribute"]], Mutability.read_only] = None
     """When an attribute is of type "complex", "subAttributes" defines a set of
     sub-attributes."""
 
-    multi_valued: Annotated[bool, Mutability.read_only]
+    multi_valued: Annotated[Optional[bool], Mutability.read_only, Required.true] = None
     """A Boolean value indicating the attribute's plurality."""
 
-    description: Annotated[str, Mutability.read_only]
+    description: Annotated[Optional[str], Mutability.read_only, Required.true] = None
     """The attribute's human-readable description."""
 
-    required: Annotated[bool, Mutability.read_only]
+    required: Annotated[Optional[bool], Mutability.read_only, Required.true] = None
     """A Boolean value that specifies whether or not the attribute is
     required."""
 
@@ -47,24 +48,32 @@ class Attribute(ComplexAttribute):
     """A collection of suggested canonical values that MAY be used (e.g.,
     "work" and "home")."""
 
-    case_exact: Annotated[bool, Mutability.read_only] = True
+    case_exact: Annotated[Optional[bool], Mutability.read_only, Required.true] = True
     """A Boolean value that specifies whether or not a string attribute is case
     sensitive."""
 
-    mutability: Annotated[Mutability, Mutability.read_only] = Mutability.read_write
+    mutability: Annotated[Mutability, Mutability.read_only, Required.true] = (
+        Mutability.read_write
+    )
     """A single keyword indicating the circumstances under which the value of
     the attribute can be (re)defined."""
 
-    returned: Annotated[Returned, Mutability.read_only] = Returned.default
+    returned: Annotated[Returned, Mutability.read_only, Required.true] = (
+        Returned.default
+    )
     """A single keyword that indicates when an attribute and associated values
     are returned in response to a GET request or in response to a PUT, POST, or
     PATCH request."""
 
-    uniqueness: Annotated[Uniqueness, Mutability.read_only] = Uniqueness.none
+    uniqueness: Annotated[Uniqueness, Mutability.read_only, Required.true] = (
+        Uniqueness.none
+    )
     """A single keyword value that specifies how the service provider enforces
     uniqueness of attribute values."""
 
-    reference_types: Annotated[Optional[List[str]], Mutability.read_only] = None
+    reference_types: Annotated[
+        Optional[List[str]], Mutability.read_only, Required.true
+    ] = None
     """A multi-valued array of JSON strings that indicate the SCIM resource
     types that may be referenced."""
 
@@ -75,15 +84,17 @@ class Schema(ComplexAttribute):
     meta: Annotated[Optional[Meta], Mutability.read_only] = None
     """A complex attribute containing resource metadata."""
 
-    id: Annotated[str, Mutability.read_only]
+    id: Annotated[Optional[str], Mutability.read_only, Required.true] = None
     """The unique URI of the schema."""
 
-    name: Annotated[Optional[str], Mutability.read_only] = None
+    name: Annotated[Optional[str], Mutability.read_only, Returned.default] = None
     """The schema's human-readable name."""
 
-    description: Annotated[Optional[str], Mutability.read_only] = None
+    description: Annotated[Optional[str], Mutability.read_only, Returned.default] = None
     """The schema's human-readable description."""
 
-    attributes: Annotated[List[Attribute], Mutability.read_only]
+    attributes: Annotated[
+        Optional[List[Attribute]], Mutability.read_only, Required.true
+    ] = None
     """A complex type that defines service provider attributes and their
     qualities via the following set of sub-attributes."""